Restructure directories: SmartContracts → move, integrate Next.js into WalGit-Frontend
