+++ WalGit-frontend/src/app/favicon.ico
