+++ WalGit-frontend/src/app/globals.css
