+++ move/sources/git_blob_object.move
