+++ move/sources/git_commit_object.move
