+++ move/sources/git_merge.move
