+++ move/sources/git_reference.move
